refactor(cms): isolate editor snapshot
Move homepage fetch and editor snapshot orchestration out of cms/index.ts so the package entry stays a thin public barrel over focused CMS modules.

Verified with pnpm --filter @anipotts/lib typecheck, pnpm --filter @anipotts/lib test, pnpm --filter @anipotts/lib build, pnpm exec prettier --check packages/lib/src/cms/index.ts packages/lib/src/cms/homepage.ts packages/lib/src/cms/snapshot.ts, git diff --check -- packages/lib/src/cms/index.ts packages/lib/src/cms/homepage.ts packages/lib/src/cms/snapshot.ts, and pnpm dlx knip --no-exit-code.
